Why do you want to learn Machine Learning?
The short answer is to find a job that pays well and that I enjoy. 

The long answer: I am extremely intellectually curious person by nature who loves to solve problems with data. Machine learning appeals to me for this reason as well as its practical business applications. 

My background is in finance and investment. I was a Portfolio Manager for several years for a very small investment fund. I developed one of the best track records of outperformance in the US during the Great Recession and the immediate years afterwards. My methodology was value-oriented; I sought investments with attractive risk-reward trade-offs. In order to find these opportunities, I’d analyze business and economic data going back decades, or in some extreme cases, even centuries. 

In spite of my track record, the investment management industry was brutal. The rise of passive investing made it more difficult to find work as an active manager, and my largest client was never happy with his returns, even when they were on Warren Buffett levels. Also unlike many people in the industry, I came from a poor family, and grew up in rural Appalachia; I never had the familial connections or money that a lot of investors have been able to leverage in order to be successful. Eventually, I had to close shop. 

The things I loved about investing was diving into the data, finding out what it means, trying to predict a future range of outcomes, and making predictions based on that. Machine learning sparks my intellectual curiosity in the same way investing does and that’s why I want to get into it. The job prospects are better and less geographically limiting (very difficult to find investment jobs in the Southeast US even in Atlanta.) 

I’ve taken courses on Udemy in Python, R, data science, and some machine learning and I’m really enjoying it. I want to take the next step.